change repository name

+++ restore.sh
@@ -5,11 +5,15 @@ if [ -z $1 ] ; then
   exit 1;
 fi
 
+work_dir=/tmp/restore`date +%s`
+if test -e $work_dir ; then
+  echo "Already exists $work_dir"
+  exit 1
+fi
 
-work_dir=/tmp/masafumi-k_`date +%s`
 rm -rf $work_dir
 mkdir -p $work_dir
-git clone https://github.com/ybiquitous/masafumi-k.git $work_dir
+git clone https://github.com/ybiquitous/dotfiles.git $work_dir
 for v in `ls -a $work_dir/HOME/` ; do
   if [ $v = '.' -o $v = '..' ] ; then
     continue
@@ -17,3 +21,5 @@ for v in `ls -a $work_dir/HOME/` ; do
   cp -rfv $work_dir/HOME/$v $1/
 done
 rm -rf $work_dir
+
+exit 0
